requirements.txt
aiofiles==25.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
altair==6.0.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
attrs==25.4.0
bidict==0.23.1
blinker==1.9.0
cachetools==6.2.4
certifi==2026.1.4
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
docutils==0.22.4
fastapi==0.128.0
frozenlist==1.8.0
gitdb==4.0.12
GitPython==3.1.46
h11==0.16.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
idna==3.11
ifaddr==0.2.0
itsdangerous==2.2.0
Jinja2==3.1.6
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
markdown2==2.5.4
MarkupSafe==3.0.3
multidict==6.7.0
narwhals==2.15.0
nicegui==3.5.0
numpy==2.4.1
oauthlib==3.3.1
orjson==3.11.5
packaging==25.0
pandas==2.3.3
pillow==12.1.0
plotly==6.5.2
propcache==0.4.1
protobuf==6.33.4
pyarrow==23.0.0
pydantic==2.12.5
pydantic_core==2.41.5
pydeck==0.9.1
Pygments==2.19.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-engineio==4.13.0
python-multipart==0.0.21
python-socketio==5.16.0
pytz==2025.2
PyYAML==6.0.3
referencing==0.37.0
requests==2.32.5
requests-oauthlib==2.0.0
rpds-py==0.30.0
simple-websocket==1.1.0
six==1.17.0
smmap==5.0.2
starlette==0.50.0
streamlit==1.53.0
tenacity==9.1.2
toml==0.10.2
tornado==6.5.4
tweepy==4.16.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
uvicorn==0.40.0
watchdog==6.0.0
watchfiles==1.1.1
websockets==16.0
wsproto==1.3.2
yarl==1.22.0
